fix bestDoM final case
List decomposition is not the correct way to check if something is a
list, since the empty list is not decomposable, but it is still a list
(duh).

This mistake led to a subtle but severe bug in the bestDoM termination
criteria, where the empty list would be wrapped in another list, and
then unwrapped again. In combination with the findall/3 used in the
pickBest bestDoM definition, this led to an infinite number of
choicepoints since each step creates a new stack level, so the findall
wouldn't terminate.

In theory, this bug should have affected any program that finals within
a pickBest within a solve block...