change trans/4 wrapper to guarantee history change
When policies generated by bestDoM are processed, not every transPr
actually leads to a new action being appended to the history. This is a
problem for golog++, since it needs to look at the head of the history
to extract the next action. However when the history is unchanged, there
is virtually no way to detect that from the ECLiPSe C++ bindings.