fix fluent nesting for set/2 action
The set/2 action has a special subf/3 handling because the fluent on the
left hand side must not be subf'ed. BUT this is only the case for the
outermost fluent symbol: Fluents nested as arguments to the outermost
fluent must still be subf'ed. The right hand side of the assignment is
however treated normally, so no change is needed there.