holds, subf: fix nesting of functionals
The previous implementation that allowed rigid & fluent functions to be
used as arguments with arbitrary nesting depth was incomplete. This
commit should now hopefully catch the remaining cases where the checks
whether something is procedure, a fluent or whatever, need to be made
independently of the supplied arguments because it happens BEFORE the
arguments get evaluated.